+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - aws_kinesis_streams: Added configurable partition keys
 - topology: Added ability to disable individual sink healthchecks
 - aws_cloudwatch_logs: Add dynamic group and stream creation
+- elasticsearch sink: Add support for custom headers and query parameters
 
 ### Changed